+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Set min date to 01/01/2021 for date picker
 - Added time to nwis line on graphs
+- Replaced 'Thresholds' and 'Reference Point' language with 'Feature(s)'
+
 
 ### Fixed